+++ rapyer/types/redis_set.py
@@ -19,6 +19,7 @@ class RedisSet(set, SpecialFieldType, Generic[T]):
     """
 
     original_type: type = set
+    LUA_SNIPPET_DIR = "redis_set"
 
     def __init__(self, *args, **kwargs):
         set.__init__(self, *args, **kwargs)
@@ -234,6 +235,11 @@ class RedisSet(set, SpecialFieldType, Generic[T]):
         members = await self.redis.smembers(self.special_key)
         if members:
             await self.client.sadd(target_special_key, *members)
+
+    def lua_save_payload(self) -> str:
+        if not self:
+            return ""
+        return json.dumps(self._dump_members(self))
 
     @classmethod
     def queue_special_loads_in_pipeline(